<delete>二つのインベントリスクリプトを作ってしまっていたので削除
