--- Assets/Z_Miyamoto/Prefabs/BoxItem.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: ab26b72496d31cd49aee9c5899615a27
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
