--- Assets/Z_Miyamoto/Prefabs/DefaultItem.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 1972bf06af353e3478d77d417cfc3fb6
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
