--- Assets/Z_Miyamoto/Scripts/Player/Inventory/Inventory2.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 30c495de93740bc42ba0dbbdb8f43309
