--- Assets/Z_Miyamoto/Scripts/Player/Inventory/PlayerInventory.asset
@@ -1,21 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: df46a43722cbe4743b687b4d24512737, type: 3}
-  m_Name: PlayerInventory
-  m_EditorClassIdentifier: 
-  Container:
-  - Item: {fileID: 11400000, guid: b30a7d74b3fd7ca47b3fe60557379bcb, type: 2}
-    Amount: 3
-  - Item: {fileID: 11400000, guid: d315b1e378a7e5a4cb736e18f7ac3f0a, type: 2}
-    Amount: 3
-  - Item: {fileID: 11400000, guid: 8f5e1938591526f4695f50b435651d8b, type: 2}
-    Amount: 3
