--- Assets/Z_Miyamoto/Scripts/Player/Inventory/PlayerInventory.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: b9892977a6525cd458638c59f518b4ab
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
